remove `removeDoubleSlash` function (let go compiler handle errors)
Removing the functionality to handle typos or similar that lead to
double slashes in the pattern like `/a//b`. The go compiler will panic
if the patter is in the new style (`go1.22`) of `GET /a//b` and the old
style will not register a route (ignore it) `/a//b`.

Tests for this behavior of go have been added.

* Order of function in muxify
* Delete tests of `removeDoubleSlash`